feat (leasson): added update lesson

+++ src/lessons/useCase/delete.useCase.ts
@@ -1,4 +1,9 @@
-import { Inject, Injectable, NotFoundException } from '@nestjs/common';
+import {
+  Inject,
+  Injectable,
+  NotFoundException,
+  UnauthorizedException,
+} from '@nestjs/common';
 import { CourseRepository } from 'src/courses/repository/course.repository';
 import { PrismaCourseRepository } from 'src/database/prisma/repository/prisma.course.repository';
 import { PrismaLessonRepository } from 'src/database/prisma/repository/prisma.lesson.repository';
@@ -11,7 +16,12 @@ export class DeleteLessonByIdUseCase {
     @Inject(PrismaCourseRepository) private courseRepository: CourseRepository,
   ) {}
 
-  async execute(id: string, courseId: string): Promise<void> {
+  async execute(id: string, courseId: string, userRole: string): Promise<void> {
+    if (['STUDENTS'].includes(userRole)) {
+      throw new UnauthorizedException(
+        'Only admins and teachers can delete lessons',
+      );
+    }
     const course = await this.courseRepository.findById(courseId);
     if (!course) throw new NotFoundException('Course not found');
     await this.lessonRepository.delete(id);
